Former-commit-id: d719920b83dc004fabf64351d62b4d25a5c5d4ff

+++ src/main/java/com/klinker/android/twitter_l/ui/drawer_activities/lists/ViewUsers.java
@@ -62,7 +62,7 @@ public class ViewUsers extends AppCompatActivity {
 
     private boolean canRefresh = true;
 
-    private int listId;
+    private long listId;
     private String listName;
 
     private long currCursor = -1;
@@ -146,7 +146,7 @@ public class ViewUsers extends AppCompatActivity {
             }
         });
 
-        listId = getIntent().getIntExtra("list_id", 0);
+        listId = getIntent().getLongExtra("list_id", 0);
 
         new GetUsers().execute();
